requirements.txt
# Django requirements
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.13
sqlparse==0.5.0
typing-extensions==4.11.0
django-mptt==0.14.0
django-mptt-admin==2.6.2

# # Image processing requirements
pillow==10.3.0
